test kit for testing json rpc servers
