+++ common/json-rpc-server-test/src/main/scala/org/enso/jsonrpc/test/JsonRpcServerTestKit.scala
@@ -1,24 +1,15 @@
-package org.enso.languageserver.websocket
-import java.nio.file.Files
-import java.util.UUID
+package org.enso.jsonrpc.test
 
 import akka.NotUsed
-import akka.actor.{ActorRef, ActorSystem, PoisonPill, Props}
+import akka.actor.{ActorRef, ActorSystem, PoisonPill}
 import akka.http.scaladsl.Http
 import akka.http.scaladsl.model.ws.{Message, TextMessage, WebSocketRequest}
 import akka.stream.OverflowStrategy
 import akka.stream.scaladsl.{Flow, Sink, Source}
 import akka.testkit.{ImplicitSender, TestKit, TestProbe}
-import cats.effect.IO
 import io.circe.Json
 import io.circe.parser.parse
-import org.enso.jsonrpc.JsonRpcServer
-import org.enso.languageserver.capability.CapabilityRouter
-import org.enso.languageserver.data.{Config, Sha3_224VersionCalculator}
-import org.enso.languageserver.filemanager.FileSystem
-import org.enso.languageserver.protocol.{JsonRpc, ServerClientControllerFactory}
-import org.enso.languageserver.text.BufferRegistry
-import org.enso.languageserver.{LanguageProtocol, LanguageServer}
+import org.enso.jsonrpc.{ClientControllerFactory, JsonRpcServer, Protocol}
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpecLike
 import org.scalatest.{Assertion, BeforeAndAfterAll, BeforeAndAfterEach}
@@ -26,7 +17,7 @@ import org.scalatest.{Assertion, BeforeAndAfterAll, BeforeAndAfterEach}
 import scala.concurrent.Await
 import scala.concurrent.duration._
 
-abstract class WebSocketServerTest
+abstract class JsonRpcServerTestKit
     extends TestKit(ActorSystem("TestSystem"))
     with ImplicitSender
     with AnyWordSpecLike
@@ -41,36 +32,16 @@ abstract class WebSocketServerTest
   val interface       = "127.0.0.1"
   var address: String = _
 
-  val testContentRoot   = Files.createTempDirectory(null)
-  val testContentRootId = UUID.randomUUID()
-  val config            = Config(Map(testContentRootId -> testContentRoot.toFile))
-
-  testContentRoot.toFile.deleteOnExit()
-
   var server: JsonRpcServer       = _
   var binding: Http.ServerBinding = _
 
+  def protocol: Protocol
+
+  def clientControllerFactory: ClientControllerFactory
+
   override def beforeEach(): Unit = {
-    val languageServer =
-      system.actorOf(
-        Props(new LanguageServer(config, new FileSystem[IO]))
-      )
-    languageServer ! LanguageProtocol.Initialize
-    val bufferRegistry =
-      system.actorOf(
-        BufferRegistry.props(languageServer)(Sha3_224VersionCalculator)
-      )
 
-    lazy val capabilityRouter =
-      system.actorOf(CapabilityRouter.props(bufferRegistry))
-
-    lazy val clientFactory = new ServerClientControllerFactory(
-      languageServer,
-      bufferRegistry,
-      capabilityRouter
-    )
-
-    server  = new JsonRpcServer(JsonRpc.protocol, clientFactory)
+    server  = new JsonRpcServer(protocol, clientControllerFactory)
     binding = Await.result(server.bind(interface, port = 0), 3.seconds)
     address = s"ws://$interface:${binding.localAddress.getPort}"
   }
